requirements.txt
asgiref==3.4.1
Django==3.1.7
django-uuslug==1.2.0
Pillow==7.2.0
python-slugify==5.0.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
